chore: remove moved files, rename folder
