--- app/react/comparision/page.tsx
@@ -1,5 +1,0 @@
-import HookComparisonsSection from "@/sections/lifecycle/HookComparisonsSection";
-
-export default function HookComparisonsSectionPage() {
-  return <HookComparisonsSection />;
-}
